moving some classes from internal to notapi
these classes offer some services to GEMOC that need to be refined in a
clear API
