+++ plugins/org.eclipse.emf.ecoretools.ale.core/META-INF/MANIFEST.MF
@@ -15,6 +15,7 @@ Export-Package: org.eclipse.emf.ecoretools.ale.core.diagnostics,
  org.eclipse.emf.ecoretools.ale.core.env.impl,
  org.eclipse.emf.ecoretools.ale.core.interpreter,
  org.eclipse.emf.ecoretools.ale.core.interpreter.impl,
+ org.eclipse.emf.ecoretools.ale.core.interpreter.notapi,
  org.eclipse.emf.ecoretools.ale.core.interpreter.services,
  org.eclipse.emf.ecoretools.ale.core.io,
  org.eclipse.emf.ecoretools.ale.core.metamodelembedding,
