+++ plugins/org.eclipse.emf.ecoretools.ale.core/src/org/eclipse/emf/ecoretools/ale/core/interpreter/services/EvalBodyService.java
@@ -38,9 +38,9 @@ import org.eclipse.emf.ecore.EPackage;
 import org.eclipse.emf.ecore.EParameter;
 import org.eclipse.emf.ecoretools.ale.core.env.IAleEnvironment;
 import org.eclipse.emf.ecoretools.ale.core.interpreter.DiagnosticLogger;
-import org.eclipse.emf.ecoretools.ale.core.interpreter.internal.EvalEnvironment;
 import org.eclipse.emf.ecoretools.ale.core.interpreter.internal.ExpressionEvaluationEngine;
 import org.eclipse.emf.ecoretools.ale.core.interpreter.internal.MethodEvaluator;
+import org.eclipse.emf.ecoretools.ale.core.interpreter.notapi.EvalEnvironment;
 import org.eclipse.emf.ecoretools.ale.core.parser.internal.AntlrAstToAleBehaviorsFactory;
 import org.eclipse.emf.ecoretools.ale.core.validation.IConvertType;
 import org.eclipse.emf.ecoretools.ale.core.validation.impl.ConvertType;
